+++ packages/dd-trace/src/opentracing/propagation/text_map.js
@@ -3,7 +3,6 @@
 const pick = require('../../../../datadog-core/src/utils/src/pick')
 const id = require('../../id')
 const DatadogSpanContext = require('../span_context')
-const OtelSpanContext = require('../../opentelemetry/span_context')
 const log = require('../../log')
 const TraceState = require('./tracestate')
 const tags = require('../../../../../ext/tags')
@@ -691,48 +690,6 @@ class TextMapPropagator {
     }
 
     return spanContext._traceId.toString(16)
-  }
-
-  static _convertOtelContextToDatadog (traceId, spanId, traceFlag, ts, meta = {}) {
-    const origin = null
-    let samplingPriority = traceFlag
-
-    ts = ts?.traceparent || null
-
-    if (ts) {
-      // Use TraceState.fromString to parse the tracestate header
-      const traceState = TraceState.fromString(ts)
-      let ddTraceStateData = null
-
-      // Extract Datadog specific trace state data
-      traceState.forVendor('dd', (state) => {
-        ddTraceStateData = state
-        return state // You might need to adjust this part based on actual logic needed
-      })
-
-      if (ddTraceStateData) {
-        // Assuming ddTraceStateData is now a Map or similar structure containing Datadog trace state data
-        // Extract values as needed, similar to the original logic
-        const samplingPriorityTs = ddTraceStateData.get('s')
-        const origin = ddTraceStateData.get('o')
-        // Convert Map to object for meta
-        const otherPropagatedTags = Object.fromEntries(ddTraceStateData.entries())
-
-        // Update meta and samplingPriority based on extracted values
-        Object.assign(meta, otherPropagatedTags)
-        samplingPriority = TextMapPropagator._getSamplingPriority(traceFlag, parseInt(samplingPriorityTs, 10), origin)
-      } else {
-        log.debug(`no dd list member in tracestate from incoming request: ${ts}`)
-      }
-    }
-
-    const spanContext = new OtelSpanContext({
-      traceId: id(traceId, 16), spanId: id(), tags: meta, parentId: id(spanId, 16)
-    })
-
-    spanContext._sampling = { priority: samplingPriority }
-    spanContext._trace = { origin }
-    return spanContext
   }
 
   static _getSamplingPriority (traceparentSampled, tracestateSamplingPriority, origin = null) {
